pyproject.toml
[project]
name = "androguard"
description = "Androguard is a full python tool to play with Android files."
authors = [{ name = "Anthony Desnos", email = "anthony@42.bzh" }]
version = "5.0.0"
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"

classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development",
    "Topic :: Utilities",
]

dependencies = ["apkparser-ag", "dex-parser"]

[project.scripts]
androguard = "androguard.main:app"

[build-system]
requires = ["setuptools>=68", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["."]
include = ["androguard", "androguard.*"]

[tool.black]
line-length = 79
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 79
